Fix typo in some messages "receive" instead of "recieve". Change some variable types for IA64 support.

+++ src/popt/poptconfig.c
@@ -58,7 +58,7 @@ int poptReadConfigFile(poptContext con, const char * fn) {
     char * file, * chptr, * end;
     char * buf, * dst;
     int fd, rc;
-    int fileLength;
+    size_t fileLength;
 
     fd = open(fn, O_RDONLY);
     if (fd < 0) {
